Create Basic Logic for Stock Picker
Created the basic logic for the stock_picker method.

This takes a value in the given array and iterates over all the other
values in the array to subtract the values.  It then creates a has
containing:
1. a two value array with the indices of the two values from the array
2. the different from the two values

The method currently iterates over all values, even values that come
before it in the array (which would not be possible with consecutive
days in real life), so the next step is to update so that it only
compares with values that come after it in the array.