Add Player Turn, Computer Check, and Move Rules
Created a player turn function that allows the player to pick a code
they would like to guess as the answer.

Added a computer check function that checks if the player's guess
matches the computer code or not.

Lastley, moved the game rules to the Game object as a function.