Add Methods to check for Red and White Pegs
Created methods under the computer object to check for what to return
to the user based on their input: red pegs, white pegs, or nothing.

I had to create multiple functions for this all at once since they are
so intertwined with one another.  The program checks for red pegs first
and then removes those choices when checking for white pegs.