Create a Start Game method
Created a method that allows the user to decided whether they want to
begin the game after given the rules.

If they want to begin they can input Y or Yes and the game will start.

If they select N or No then it will display a message and exit the
program.