Add Method for Player to Choose Game Mode
Created a method for the player to choose whether they want to be the
code guesser or the code maker.

Depending on their choice this will either take them through the
existing program of being the guesser or through a branch that currently
does not do anything.

The next step is to build out the logic so that the player can be the
code maker and have the computer guess.