Create full game loop
Created the full game loop in the main file that will determine if the
player successfully guesses the word or if they run out of chances.

I also created a function to check if the user had already incorrectly
guessed a letter previously and for it to not count against them.