Create Game State Class
Created a class that will allow the user to svae the current state of
the game and began building out the logic to use it in the game loop.

This will be further built out to create a save file that can then be
called at the beginning of other games if you would like to load it.