Create Linked List Data Structure
Created a Linked List Data Structure with various methods that allow you
to add to/remove from the list as well as learn additional details
about the list.

The list will be in the order you create it all linking in one direction
and the order will depend on whether you append, prepend, insert, or
remove data.