Implement Increase Capacity Functionality
Created a method that checks if the capacity of the hash map needs to be
increased or not based on the load factor.

If it does need to be increased, then it recreates the hash map with
double the load factor and spreads the original entries through the
new hash map evenly.