Create Level Order Method
This method traverses the binary tree in a breadth first level order and
creates an array of the nodes in this order.

The user can then either generate a black against this array and
manipulate it or enter the method without a block and it will return
just the array in breadth order.