Create Pre-,In-, and Postorder Methods
These methods traverse the tree in a depth first manner and do so
according to their type (pre, in, or post).

Just like the level-order traversal, if a block is given it will yield
each node to the block.  Otherwise, it will just give an array of the
values for each node in the order of the specified traversal.